remove unordered leaf feature, optimize prepend/append insert paths, improve SIMD UTs, optimize lookups to use pointer comparison (which we'll get from the paired hashtable), use -1 for no-result rank queries, simplify node header

+++ src/skiplist_ordered_index.c
@@ -23,8 +23,8 @@ static OrderedIndexItem *skiplistGetByRank(OrderedIndex *idx, unsigned long rank
     return (OrderedIndexItem *)zslGetElementByRank((zskiplist *)idx, rank);
 }
 
-static unsigned long skiplistGetRank(OrderedIndex *idx, const OrderedIndexItem *node) {
-    return zslGetRank((zskiplist *)idx, (const zskiplistNode *)node);
+static long skiplistGetRank(OrderedIndex *idx, const OrderedIndexItem *node) {
+    return (long)zslGetRank((zskiplist *)idx, (const zskiplistNode *)node);
 }
 
 static unsigned long skiplistLength(OrderedIndex *idx) {
